hopefully fix call to start
on windows 10, the title seems to be optional but on windows 11 it seems
to be required. However setting it to something meaningful confuses the
start command on windows 10 to think the title is the command. Setting
it to an empty string seems to be the only cross version compatible way.

WTF Microsoft?